feat: Ext 추가 및 TabBar 스타일 변경
