+++ RF/Extensions/CALayer+Ext.swift
@@ -1,0 +1,25 @@
+//
+//  CALayer+Ext.swift
+//  RF
+//
+//  Created by 이정동 on 2023/07/09.
+//
+
+import Foundation
+import UIKit
+
+extension CALayer {
+    // Sketch 스타일의 그림자를 생성하는 유틸리티 함수
+    func applyShadow(
+        color: UIColor = .black,
+        alpha: Float = 0.5,
+        x: CGFloat = 0,
+        y: CGFloat = 2,
+        blur: CGFloat = 4
+    ) {
+        shadowColor = color.cgColor
+        shadowOpacity = alpha
+        shadowOffset = CGSize(width: x, height: y)
+        shadowRadius = blur / 2.0
+    }
+}
