+++ RF/Extensions/UITabBar+Ext.swift
@@ -1,0 +1,18 @@
+//
+//  UITabBar+Ext.swift
+//  RF
+//
+//  Created by 이정동 on 2023/07/09.
+//
+
+import Foundation
+import UIKit
+
+extension UITabBar {
+    // 기본 그림자 스타일을 초기화해야 커스텀 스타일을 적용할 수 있다.
+    static func clearShadow() {
+        UITabBar.appearance().shadowImage = UIImage()
+        UITabBar.appearance().backgroundImage = UIImage()
+        UITabBar.appearance().backgroundColor = UIColor.white
+    }
+}
