+++ RF/Views/Meeting/MeetingViewController.swift
@@ -209,30 +209,4 @@ extension MeetingViewController: ClickedButton {
 
 
 // 파일 옮기기
-extension UIImage {
-    // MARK: 이미지 크기 재배치 하는 함수
-    func resize(newWidth: CGFloat) -> UIImage {
-        let scale = newWidth / self.size.width
-        let newHeight = self.size.height * scale
-        
-        let size = CGSize(width: newWidth, height: newHeight)
-        let render = UIGraphicsImageRenderer(size: size)
-        let renderImage = render.image { context in
-            self.draw(in: CGRect(origin: .zero, size: size))
-        }
-        return renderImage
-    }
-    
-    // MARK: 이미지 크기 (강제)재배치 하는 함수
-    func resize(newWidth: CGFloat, newHeight: CGFloat) -> UIImage {
-        
-        let size = CGSize(width: newWidth, height: newHeight)
-        let render = UIGraphicsImageRenderer(size: size)
-        let renderImage = render.image { context in
-            self.draw(in: CGRect(origin: .zero, size: size))
-        }
-        return renderImage
-    }
-    
-}
